requirements.txt
Pillow
python-barcode